fix: embed SPA frontend into Rust backend binary
- Add rust-embed dependency to Cargo.toml and board crate
- Implement SPA fallback handler using rust-embed to serve
  Next.js static assets from compiled binary
- Replace multi-container deployment with single-container
  approach where Rust backend serves both API and static SPA
- Update Dockerfile to build Next.js frontend in separate stage
  and copy assets for embedding during Rust compilation
- Refactor chunked upload and room handlers to use centralized
  room name validation via RoomNameValidator
- Remove nginx gateway and standalone frontend service from
  docker-compose.yml
- Update dependency versions in Cargo.lock including hyper 1.9.0
  to 1.10.0, serde_json 1.0.149 to 1.0.150, and others
