+++ docker-compose.yml
@@ -1,61 +1,19 @@
 # ============================================================================
 # Docker Compose configuration for Elizabeth
 # ============================================================================
-# File sharing and collaboration platform
+# Single-container deployment: Rust binary with embedded SPA frontend
+# One port serves both the API backend and the static SPA frontend.
 # ============================================================================
 
 services:
   # ==========================================================================
-  # Gateway (Single public entrypoint)
+  # Backend Service (Rust + Axum + Embedded Next.js SPA)
   # ==========================================================================
-  # Expose ONLY this service to the host. It proxies:
-  # - /            -> frontend (Next.js)
-  # - /api/v1/*    -> backend  (Axum)
-  # - /ws          -> backend  (WebSocket)
-  #
-  # This prevents browsers from reaching the backend directly, avoiding mixed
-  # content / CORS issues and keeping the backend off the public network.
-  gateway:
-    image: nginx:stable-alpine
-    container_name: elizabeth-gateway
-    restart: unless-stopped
-    init: true
-    security_opt:
-      - no-new-privileges:true
-    read_only: true
-    tmpfs:
-      - /tmp
-      - /var/cache/nginx
-      - /var/run
-    ports:
-      - "${FRONTEND_PORT:-4001}:4001"
-    volumes:
-      - ./docker/gateway/nginx.conf:/etc/nginx/conf.d/default.conf:ro
-    networks:
-      - elizabeth-network
-    depends_on:
-      backend:
-        condition: service_healthy
-      frontend:
-        condition: service_healthy
-    healthcheck:
-      test:
-        [
-          "CMD",
-          "wget",
-          "--no-verbose",
-          "--tries=1",
-          "--spider",
-          "http://127.0.0.1:4001/",
-        ]
-      interval: 30s
-      timeout: 3s
-      retries: 3
-      start_period: 10s
-
-  # ==========================================================================
-  # Backend Service (Rust + Axum)
-  # ==========================================================================
+  # Single Rust binary serves:
+  # - /api/v1/*       -> Backend API (Axum handlers)
+  # - /api/v1/ws      -> WebSocket (real-time sync)
+  # - /_next/*, etc.  -> Embedded SPA static assets (rust-embed)
+  # - /*              -> SPA HTML fallback (index.html / _.html)
   backend:
     build:
       context: .
@@ -63,7 +21,7 @@ services:
       extra_hosts:
         - "index.crates.io=${CRATES_IO_INDEX_IPV4:-151.101.89.91}"
         - "static.crates.io=${CRATES_IO_STATIC_IPV4:-151.101.90.137}"
-    container_name: elizabeth-backend
+    container_name: elizabeth
     restart: unless-stopped
     init: true
     security_opt:
@@ -73,6 +31,8 @@ services:
     read_only: true
     tmpfs:
       - /tmp
+    ports:
+      - "${PORT:-4092}:4092"
     volumes:
       # Data persistence
       - ${ELIZABETH_DATA_DIR:-./docker/backend/data}:/app/data
@@ -94,68 +54,9 @@ services:
       DB_MIN_CONNECTIONS: ${DB_MIN_CONNECTIONS:-1}
       # Room configuration
       ELIZABETH__APP__ROOM__SHARE_DISABLED_LOCK_DURATION: ${ROOM_SHARE_DISABLED_LOCK_DURATION:-3600}
-    expose:
-      - "4092"
-    networks:
-      - elizabeth-network
     healthcheck:
       test: ["CMD", "curl", "-f", "http://127.0.0.1:4092/api/v1/health"]
       interval: 30s
       timeout: 10s
       retries: 3
       start_period: 10s
-
-  # ==========================================================================
-  # Frontend Service (Next.js 16)
-  # ==========================================================================
-  frontend:
-    build:
-      context: .
-      dockerfile: Dockerfile.frontend
-      args:
-        NEXT_PUBLIC_API_URL: ${NEXT_PUBLIC_API_URL:-/api/v1}
-        NEXT_PUBLIC_APP_URL: ${NEXT_PUBLIC_APP_URL:-http://localhost:4001}
-        INTERNAL_API_URL: ${INTERNAL_API_URL:-http://elizabeth-backend:4092/api/v1}
-    container_name: elizabeth-frontend
-    restart: unless-stopped
-    init: true
-    security_opt:
-      - no-new-privileges:true
-    cap_drop:
-      - ALL
-    environment:
-      # Public variables (client-side)
-      NEXT_PUBLIC_API_URL: ${NEXT_PUBLIC_API_URL:-/api/v1}
-      NEXT_PUBLIC_APP_URL: ${NEXT_PUBLIC_APP_URL:-http://localhost:4001}
-      # Internal variables (server-side)
-      INTERNAL_API_URL: ${INTERNAL_API_URL:-http://elizabeth-backend:4092/api/v1}
-      # Node.js
-      NODE_ENV: production
-      PORT: 4001
-      HOSTNAME: 0.0.0.0
-    networks:
-      - elizabeth-network
-    depends_on:
-      backend:
-        condition: service_healthy
-    healthcheck:
-      test:
-        [
-          "CMD",
-          "wget",
-          "--no-verbose",
-          "--tries=1",
-          "--spider",
-          "http://127.0.0.1:4001/",
-        ]
-      interval: 30s
-      timeout: 3s
-      retries: 3
-      start_period: 15s
-
-# ============================================================================
-# Networks
-# ============================================================================
-networks:
-  elizabeth-network:
-    driver: bridge
