+++ web/app/page.tsx
@@ -8,8 +8,7 @@ export const metadata: Metadata = {
   generator: "unic",
 };
 
-export const dynamic = "force-dynamic";
-export const revalidate = 0;
+
 
 export default function Page() {
   return <HomePage />;
