+++ web/e2e/tests/file-upload-correct.spec.ts
@@ -2,7 +2,7 @@ import { test, expect } from "@playwright/test";
 import * as fs from "fs";
 import { RoomPage } from "../page-objects/room-page";
 
-const BASE_URL = "http://localhost:4001";
+const BASE_URL = "http://localhost:4092";
 const API_BASE = "http://localhost:4092/api/v1";
 const TEST_ROOM = "file-upload-correct";
 const TEST_ROOM_URL = `${BASE_URL}/${TEST_ROOM}`;
