+++ web/e2e/tests/room-lifecycle-limits.spec.ts
@@ -2,7 +2,7 @@
 import { expect, test } from "@playwright/test";
 import { RoomPage } from "../page-objects/room-page";
 
-const BASE_URL = "http://localhost:4001";
+const BASE_URL = "http://localhost:4092";
 const API_BASE_URL = process.env.API_BASE_URL || "http://localhost:4092/api/v1";
 
 async function ensureRoomExists(roomName: string) {
