+++ web/e2e/tests/sample-room-tests.spec.ts
@@ -6,7 +6,7 @@
 import { expect, type Page, test } from "@playwright/test";
 import { RoomPage } from "../page-objects/room-page";
 
-const BASE_URL = "http://localhost:4001";
+const BASE_URL = "http://localhost:4092";
 const API_BASE_URL = process.env.API_BASE_URL || "http://localhost:4092/api/v1";
 const TOKEN_STORAGE_KEY = "elizabeth_tokens";
