+++ web/next.config.mjs
@@ -1,4 +1,6 @@
 /** @type {import('next').NextConfig} */
+const isProd = process.env.NODE_ENV === "production";
+
 const nextConfig = {
   typescript: {
     ignoreBuildErrors: false,
@@ -7,41 +9,45 @@ const nextConfig = {
     unoptimized: true,
   },
   allowedDevOrigins: ["local-origin.dev", "*.local-origin.dev"],
-  // Enable standalone output for Docker deployment
-  output: "standalone",
 
-  async headers() {
-    return [
-      {
-        source: "/",
-        headers: [
-          {
-            key: "Cache-Control",
-            value: "no-store, max-age=0, must-revalidate",
-          },
-        ],
-      },
-    ];
-  },
+  // 生产模式：静态导出 (web/out/)，由 Rust 二进制 rust-embed 打包并 serve
+  // 开发模式：standalone，保留 Next.js dev server + API rewrites 代理
+  output: isProd ? "export" : "standalone",
 
-  // API rewrites for backend proxy - Next.js 16 compatible version
-  async rewrites() {
-    const target = process.env.INTERNAL_API_URL;
+  // 仅本地开发时：将 /api/v1/* 代理到后端（4092 端口）
+  ...(isProd
+    ? {}
+    : {
+        async headers() {
+          return [
+            {
+              source: "/",
+              headers: [
+                {
+                  key: "Cache-Control",
+                  value: "no-store, max-age=0, must-revalidate",
+                },
+              ],
+            },
+          ];
+        },
 
-    console.log('Next.js: INTERNAL_API_URL from env:', target);
+        async rewrites() {
+          const target =
+            process.env.INTERNAL_API_URL ||
+            process.env.NEXT_PUBLIC_BACKEND_URL ||
+            "http://127.0.0.1:4092/api/v1";
 
-    if (!target) {
-      console.warn('WARNING: INTERNAL_API_URL is not set. API rewrites will not work.');
-      return [];
-    }
+          console.log("Next.js: API proxy target:", target);
 
-    return [
-      {
-        source: '/api/v1/:path*',
-        destination: `${target}/:path*`,
-      },
-    ];
-  },
+          return [
+            {
+              source: "/api/v1/:path*",
+              destination: `${target}/:path*`,
+            },
+          ];
+        },
+      }),
 };
 
 export default nextConfig;
